update permission join, edit, follow using Police

+++ resources/views/user/home/new/index.blade.php
@@ -26,11 +26,11 @@
                             @endif
                             <br/>
                         <div class="col-md-12 col-sm-12 col-xs-12">
-                              
+
                             <div class="x_panel">
-                                   
+
                                 <div class="x_title">
-                                        
+
 
                                     <h2> Newest Trips <small>Sessions</small></h2>
                                     <ul class="nav navbar-right panel_toolbox">
@@ -63,7 +63,7 @@
                                     </thead>
 
                                     <tbody>
-                                        
+
                                             @foreach($trip as $t)
                                             <tr>
                                                 <td>{{$loop->iteration}}</td>
@@ -73,22 +73,28 @@
                                                 <td>{{$t->people_number}}</td>
                                                 <td>{{$t->status}}</td>
                                                 <td><a href="{{url('user/trip/'.$t->id)}}"><button class="btn btn-info btn-sm"><i class="fa fa-eye"></i> View</button></a>
-                                                    @if(Auth::user()->id <> $t->owner_id)
-                                                        @if( Auth::user()->checkFollow(Auth::user()->id, $t->id) )
-                                                        <a href="{{url('user/trip/follow/follow/'.$t->id)}}"><button class="btn btn-success btn-sm"><i class="fa fa-star"></i> Flow</button></a>
-                                                        @else
-                                                        <a href="{{url('user/trip/follow/unfollow/'.$t->id)}}"><button class="btn btn-warning btn-sm"><i class="fa fa-star"></i> Unfollow</button></a>
-                                                        @endif
+                                                    @can('ablePlan', $t)
+                                                      @cannot('updateTrip', $t)
 
-                                                        @if(Auth::user()->checkVerify(Auth::user()->id, $t->id) && Auth::user()->checkJoin(Auth::user()->id, $t->id))
-                                                            <a href="{{url('user/trip/verify/verify/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Verify</button></a>
-                                                        @elseif(!Auth::user()->checkVerify(Auth::user()->id, $t->id) && Auth::user()->checkJoin(Auth::user()->id, $t->id)) 
-                                                            <a href="{{url('user/trip/verify/unverify/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Unverify</button></a>
-                                                        @elseif(!Auth::user()->checkJoin(Auth::user()->id, $t->id))
-                                                            <a href="{{url('user/trip/join/unjoin/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Unjoin</button></a>
-                                                        @endif     
-                                                    @else 
+                                                        @can('follow', $t)
+                                                          <a href="{{ url('/user/trip/follow/unfollow/' . $t->id) }}"><button class="btn btn-warning btn-sm"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Unfollow</button></a>
+                                                        @else
+                                                          <a href="{{ url('/user/trip/follow/follow/' . $t->id) }}"><button class="btn btn-success btn-sm"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Follow</button></a>
+                                                        @endcan
+                                                        @can('joinAble', $t)
+                                                          <a href="{{url('user/trip/verify/verify/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Join</button></a>
+                                                        @endcan
+                                                        @can('join', $t)
+                                                          <a href="{{url('user/trip/join/unjoin/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Unjoin</button></a>
+                                                        @endcan
+                                                        @can('verify', $t)
+                                                          <a href="{{url('user/trip/verify/unverify/'.$t->id)}}"><button class="btn btn-danger btn-sm"><i class="fa fa-group"></i> Unverify</button></a>
+                                                        @endcan
+                                                      @endcannot
+                                                @can('updateTrip', $t)
+
                                                     <a href="{{ url('/user/trip/' . $t->id . '/edit') }}" title="Edit trip"><button class="btn btn-primary btn-sm"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Edit</button></a>
+
                                                     {!! Form::open([
                                                         'method'=>'DELETE',
                                                         'url' => ['/user/trip', $t->id],
@@ -101,7 +107,8 @@
                                                                 'onclick'=>'return confirm("Confirm delete?")'
                                                         ]) !!}
                                                     {!! Form::close() !!}
-                                                    @endif
+                                                    @endcan
+                                                @endcan
                                                 </td>
                                             </tr>
                                             @endforeach
